refactor(auth): move packages under credentials
