+++ auth/credentials/idtoken/idtoken_test.go
@@ -29,7 +29,7 @@ import (
 
 func TestNewCredentials_ServiceAccount(t *testing.T) {
 	wantTok, _ := createRS256JWT(t)
-	b, err := os.ReadFile("../internal/testdata/sa.json")
+	b, err := os.ReadFile("../../internal/testdata/sa.json")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -87,7 +87,7 @@ func TestNewCredentials_ImpersonatedServiceAccount(t *testing.T) {
 	}
 	creds, err := NewCredentials(&Options{
 		Audience:        "aud",
-		CredentialsFile: "../internal/testdata/imp.json",
+		CredentialsFile: "../../internal/testdata/imp.json",
 		CustomClaims: map[string]interface{}{
 			"foo": "bar",
 		},
