+++ auth/credentials/impersonate/integration_test.go
@@ -26,8 +26,8 @@ import (
 
 	"cloud.google.com/go/auth"
 	"cloud.google.com/go/auth/credentials"
-	"cloud.google.com/go/auth/idtoken"
-	"cloud.google.com/go/auth/impersonate"
+	"cloud.google.com/go/auth/credentials/idtoken"
+	"cloud.google.com/go/auth/credentials/impersonate"
 	"cloud.google.com/go/auth/internal/testutil"
 	"cloud.google.com/go/auth/internal/testutil/testgcs"
 )
